+++ src/components/Export/ExportPoll.vue
@@ -29,7 +29,6 @@ import { Answer, AnswerSymbol, useVotesStore } from '../../stores/votes.ts'
 import { Option, useOptionsStore } from '../../stores/options.ts'
 import { AxiosError } from '@nextcloud/axios'
 import { DateTime, Interval } from 'luxon'
-import { useSessionStore } from '../../stores/session.ts'
 
 enum ArrayStyle {
 	Symbols = 'symbols',
@@ -48,7 +47,6 @@ const route = useRoute()
 const pollStore = usePollStore()
 const votesStore = useVotesStore()
 const optionsStore = useOptionsStore()
-const sessionStore = useSessionStore()
 
 const regex = /[:\\/?*[\]]/g
 
@@ -211,9 +209,9 @@ function getIntervalRaw(option: Option): string {
  * @param option - option
  */
 function getFromFormatted(option: Option): string {
-	return DateTime.fromSeconds(option.timestamp)
-		.setLocale(sessionStore.currentUser.languageCodeIntl)
-		.toLocaleString(DateTime.DATETIME_MED_WITH_WEEKDAY)
+	return DateTime.fromSeconds(option.timestamp).toLocaleString(
+		DateTime.DATETIME_MED_WITH_WEEKDAY,
+	)
 }
 
 /**
@@ -222,7 +220,6 @@ function getFromFormatted(option: Option): string {
  */
 function getToFormatted(option: Option): string {
 	return DateTime.fromSeconds(option.timestamp)
-		.setLocale(sessionStore.currentUser.languageCodeIntl)
 		.plus({ seconds: option.duration })
 		.toLocaleString(DateTime.DATETIME_MED_WITH_WEEKDAY)
 }